Rename view to tui
